Temporary fix for -finstrument.
In the vanilla build, the undefined symbol
_ZNSt7__cxx1118basic_stringstreamIcSt11char_traitsIcESaIcEED1Ev
exists in libspindleflux.so.0.0.0 and causes no issues.
Demangled this is
        std::__cxx11::basic_stringstream<
                char, std::char_traits<char>,
                std::allocator<char> >
                ::~basic_stringstream()

When compiled with -finstrument, the constructor is also present
and undefined, which causes a runtime error.
_ZNSt7__cxx1118basic_stringstreamIcSt11char_traitsIcESaIcEEC1Ev
        std::__cxx11::basic_stringstream<char,
                std::char_traits<char>,
                std::allocator<char> >
                ::basic_stringstream()

        0.161s: flux-shell[0]: ERROR: dlopen:
        /p/vast1/rountree/tuolumne/install/Spindle/lib/
        libspindleflux.so: undefined symbol:
        _ZNSt7__cxx1118basic_stringstreamIc

(Flux truncates the error message.)

This PR changes the code to use a stringstream constructor that
takes a single parameter:  ios_base::in | ios_base::out.

Thanks to Matt for the fix.